+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add exporting of report in Markdown format
 - Add browser access to restarting the API
+- Display tooltip for clipboard and download buttons in export menu
 
 ### Other